Fixed Summation problem for float16, float32, float64
(floatXX = float16 + floatXX)

+++ serie2/Sum.py
@@ -1,11 +1,14 @@
 # Authors: Tom Lambert (lambertt) and Yuuma Odaka-Falush (odafaluy)
 
 
+import numpy
+
+
 def sum_indices(addends):
     """
     Sums the given addends, starting by the first number and following the list
     """
-    result = 0
+    result = numpy.float16(0)
     for addend in addends:
         result += addend
     return result
@@ -15,7 +18,7 @@ def sum_ordered(addends):
     """
     Sums the given addends, starting by the smallest and following the natural order.
     """
-    result = 0
+    result = numpy.float16(0)
     for addend in sorted(addends):
         result += addend
     return result
@@ -26,8 +29,8 @@ def sum_ordered_grouped_by_sign(addends):
     Sums the given addends; All negative and all positive values are added separately
     and will be added together in the last step.
     """
-    pos = 0
-    neg = 0
+    pos = numpy.float16(0)
+    neg = numpy.float16(0)
     for addend in sorted(addends):
         if addend > 0:
             pos += addend
